E1/requirements.txt
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
nltk>=3.8.0
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.0
plotly>=5.14.0
jupyter>=1.0.0
ipywidgets>=8.0.0
spacy>=3.5.0
datasets>=2.14.0
python-Levenshtein>=0.21.0
accelerate
evaluate>=0.4.0
zstandard>=0.21.0
